requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
joblib
pyttsx3
streamlit
fpdf
mysql-connector-python